requirements.txt
faker
boto3 
python-dotenv
awscli
